+++ DSP2803x_examples_ccsv5/i2c_eeprom/User/User_Prototypes.h
@@ -40,8 +40,8 @@ void Process_System_Status(void);
 void Calculate_SOH(void);
 
 void Calculate_SOC();
-float interpolate_segment(float x0, float y0, float x1, float y1, float x);
-float interpolate_table_1d(struct table_1d *table, float x);
+long interpolate_segment(long x0, long y0, long x1, long y1, long x);
+long interpolate_table_1d(struct table_1d *table, long x);
 
 //void Calibrate_Current(void);
